requirements.txt
snowflake-connector-python
pandas
plotly-express
matplotlib
seaborn